Try to handle local dev network and 'real' ones
The idea is that the front-end works whether it's
pointing at a real blockchain (such as Rinkeby),
or a local ganache-cli instance (with persistent
build artefacts enabled).

This is tricky, because useEthers() reports the
chainId as 1337, but brownie uses "dev" in the
map.json file instead.

This code seems to work, but I get a weird issue
when using Metamask, targeting localhost:8545 - 
when I first try to connect to the application,
metamask asks me to approve a contract deployment
transaction from the deployer account. I *think*
this is something to do with nonces, where
metamask things there's an outstanding transaction
that needs to be approved. This would make sense,
because I'm resetting my local blockchain when I
restart it, which would account for nonces going
out of sync.

It doesn't happen on Rinkeby, so I'm just going to
live with it, for now.