Add instructions for fixing a usedapp issue
This project uses both useDapp and ethersproject,
but useDapp depends on ethersproject, so it 
installs its own version of ethersproject in its
own node_modules directory.

This was causing a problem a `Contract` parameter
was being passed to a function, which threw an
error because the type was not compatible with its
signature.

See: https://github.com/TrueFiEng/useDApp/issues/263#issuecomment-961158657